requirements.txt
alembic==1.1.0
blinker==1.4
cffi==1.12.3
Click==7.0
elasticsearch==7.0.5
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.6
Werkzeug==0.15.5
WTForms==2.2.1
